Fix x86 compilation error in Convert plugin

+++ examples/converter/source/conversionPanel.d
@@ -170,7 +170,7 @@ public:
 		}
 
 	nothrow @nogc
-	void setValueExcept(int exceptID, size_t value)
+	void setValueExcept(int exceptID, ulong value)
 
 		do
 		{
